Fix for CPU-spinning on device disconnection
If a serial port device was disconnected, we got stuck in read loop burning CPU cycles and having a unusable device.

On disconnect (at least on FreeBSD), select() will return 1 indicating
FD ready to read. The subsequent read() would however return 0. If this
happens, break out of the read loop and re-init the connection.